backend/api_backend/requirements.txt
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-Cors==5.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.1.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
